Integrate dependency injection and enhance logging in game initialization
- Added go.uber.org/dig for dependency injection to manage game components.
- Refactored main function to initialize logger and provide game configuration and instance through the DI container.
- Updated NewGimlarGame function to accept logger and config parameters, improving error handling and logging.
- Enhanced player image loading and error logging with zap logger.
- Introduced new methods in the player package for better coordinate calculations and input handling.

+++ game/player.go
@@ -222,8 +222,8 @@ func (player *Player) getRotatedSprite() *ebiten.Image {
 // calculatePosition calculates the new position based on the current viewAngle
 func (player *Player) calculatePosition() resolv.Vector {
 	return resolv.Vector{
-		X: center.X + radius*math.Cos(player.viewAngle),
-		Y: center.Y - radius*math.Sin(player.viewAngle),
+		X: float64(center.X) + radius*math.Cos(player.viewAngle),
+		Y: float64(center.Y) - radius*math.Sin(player.viewAngle),
 	}
 }
